Move datetime tests into the new module
